Limit username characterset
This will only effect the registration of new users. Users that already
have a username that would be against the restrictions are not effected.

We use regex here because it's handsome.

NOTE: The error we report on bad username becomes somewhat cluttered but
I think it's fine since it's going to be reported for the user only anyways.